+++ .storybook/main.js
@@ -21,7 +21,6 @@ module.exports = {
   addons: [
     '@storybook/addon-a11y',
     '@storybook/addon-actions',
-    '@storybook/addon-knobs',
     '@storybook/addon-viewport',
   ],
 };