+++ packages/bpk-component-switch/src/BpkSwitch.js
@@ -41,10 +41,11 @@ export type Props = {
 
 const BpkSwitch = (props: Props) => {
   const { className, label, small, type, ...rest } = props;
+  const resolvedType = type || SWITCH_TYPES.primary;
 
   const switchClassNames = getClassName(
     'bpk-switch__switch',
-    `bpk-switch__switch--${type}`,
+    `bpk-switch__switch--${resolvedType}`,
     small && 'bpk-switch__switch--small',
   );
 
@@ -77,7 +78,7 @@ BpkSwitch.propTypes = {
 
 BpkSwitch.defaultProps = {
   className: null,
-  type: SWITCH_TYPES.primary,
+  type: null,
   small: false,
 };